implement getPosts from db

+++ server/controllers/post.controller.js
@@ -1,12 +1,36 @@
-import Post from "../models/post.model.js"
+import Post from "../models/post.model.js";
+import User from "../models/user.model.js"
 
 
 export const createPost = async (req, res) => {
 
-    const newPost = new Post(req.body);
+    const clerkUserId = req.auth.userId;
+
+    if (!clerkUserId) {
+        return res.status(401).json("Not authenticated!");
+    }
+
+    const user = await User.findOne({ clerkUserId });
+
+    if (!user) {
+        return res.status(404).json("User not found!");
+    }
+
+    let slug = req.body.title.replace(/ /g, "-").toLowerCase();
+
+    let existingPost = await Post.findOne({ slug });
+
+    let counter = 2;
+
+    while (existingPost) {
+        slug = `${slug}-${counter}`;
+        existingPost = await Post.findOne({ slug });
+        counter++;
+    }
+
+    const newPost = new Post({ user: user._id, slug, ...req.body });
 
     const post = await newPost.save();
-
     res.status(200).json(post);
 
 }
@@ -15,6 +39,8 @@ export const createPost = async (req, res) => {
 export const getPosts = async (req, res) => {
 
     const posts = await Post.find();
+
+    if(!posts) return res.status(502).json('No created post!');
     
     res.status(200).json(posts)
 
@@ -32,8 +58,30 @@ export const getPost = async (req, res) => {
 
 export const deletePost = async (req, res) => {
 
-    await Post.findByIdAndDelete(req.params.id)
-    
-    res.status(200).json({message: 'Post has been deleted'});
+    const clerkUserId = req.auth.userId;
+
+    if (!clerkUserId) {
+        return res.status(401).json("Not authenticated!");
+    }
+
+    const role = req.auth.sessionClaims?.metadata?.role || "user";
+
+    if (role === "admin") {
+        await Post.findByIdAndDelete(req.params.id);
+        return res.status(200).json("Post has been deleted");
+    }
+
+    const user = await User.findOne({ clerkUserId });
+
+    const deletedPost = await Post.findOneAndDelete({
+        _id: req.params.id,
+        user: user._id,
+    });
+
+    if (!deletedPost) {
+        return res.status(403).json("You can delete only your posts!");
+    }
+
+    res.status(200).json("Post has been deleted");
 }
